pulling out reused test functionality into seperate files to be used later
